In Progress
.

+++ TestSutraUsers/src/main/java/com/tathvatech/workstation/common/WorkstationQuery.java
@@ -5,8 +5,11 @@ import java.util.Date;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.tathvatech.user.OID.SiteOID;
 import com.tathvatech.user.OID.TSBeanBase;
 import com.tathvatech.user.OID.WorkstationOID;
+import com.tathvatech.user.entity.Site;
+import com.tathvatech.user.entity.SiteCache;
 import com.tathvatech.user.entity.User;
 import com.tathvatech.workstation.inf.WorkstationOrderComparable;
 
@@ -117,7 +120,7 @@ public class WorkstationQuery extends TSBeanBase implements WorkstationOrderComp
 	
 	public WorkstationOID getOID()
 	{
-		return new WorkstationOID(Math.toIntExact(pk), getDisplayString());
+		return new WorkstationOID((int) pk, getDisplayString());
 	}
 	
 	@Override
@@ -136,7 +139,7 @@ public class WorkstationQuery extends TSBeanBase implements WorkstationOrderComp
 	@Override
 	public int hashCode()
 	{
-		return Math.toIntExact(pk);
+		return (int) pk;
 	}
 	@Override
 	public String toString() {
